Last Push -JE

+++ src/TeamCComponents/backend/teamcapplication/src/main/java/com/teamcid/teamcapplication/repository/courseRepository.java
@@ -1,9 +1,8 @@
+// Source code is decompiled from a .class file using FernFlower decompiler.
 package com.teamcid.teamcapplication.repository;
 
+import com.teamcid.teamcapplication.model.course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.teamcid.teamcapplication.model.course;
-
-public interface courseRepository extends JpaRepository<course, Integer>{
-
+public interface courseRepository extends JpaRepository<course, Long> {
 }
